requirements.txt
yfinance==0.2.33
pandas==2.2.2
plotly==5.22.0
python-dotenv==1.0.1